elasticarray: fix unsigned arithmetic wrap around bug
In the case of modulo wrapping of nsize * 4, the else if block could be
taken even though EA->alloc is not really greater than nsize * 4.  This
can lead to nalloc becoming smaller than nsize, which will result in the
buffer to be reallocated to be smaller than the desired nsize.