Actions: allow manual triggering
I tried to get Actions to run on contributor's PRs back in:
    2023-09-19 Actions: don't ignore if '/' in branch name
    748cb93c25e57b1d6c071d66c17816bb6f94b9f8
but it didn't work.

After reading the docs some more, my theory is that the since the
previous yaml ran "on a push to any branch", that meant "on a push to
any branch in the Tarsnap/tarsnap repository", which didn't apply to
third parties, since they were pushing to their own fork.

Adding "workflow_dispatch" should allow us to manualy trigger a check
for third-party PRs.