multitape_open: disabling chunkification if dryrun is 2
Disabling chunkification also stops any file or header data from
reaching the chunk layer.

Metadata and metaindex bytes are unaffected by this change; those are
still passed on.  This is not a deliberate choice; it's just a bit more
awkward to stop those bytes from going to the chunk layer, and since we
disabled the combination of --print-stats with --dry-run-metadata in
bsdtar.c, it doesn't matter if those bytes are passed on.  There's no
noticeable performance impact of metadata and metaindex bytes.