Add --null-output for --list-archives, instead of --null
This should have been part of:
    2023-08-30 Allow --list-archives --null
    d21dff3f1fc598b0a52008597293447dc4dbd626

The --null option was previously used for reading input with -T or -X.
When we extended it to apply to --list-archives as well, we were using
it to modify the *output* behaviour.

That wasn't ideal, but we could have lived with it, because there isn't
any input to --list-archives.  But suppose that we wanted to allow -T
with --list-archives --hashes (i.e. provide a filename which contains
hashes which we want to print)?  The user would be stuck with either
having NUL-separated both input and output, or having newlines for both.

Adding --null-output allows the user to have different separators for
the input and output.  (We can think of "--null" as "--null-input", but
we're not renaming it for historical / bsdtar-compatibility reasons.)