tree.c: Remove dead code
Remove #if 0 tree_dump debug code which has been unused for 20 years.

The field (struct tree).current is now set but never used (except in
a conditional which only involves setting that same field), and the
field (struct tree_entry).parent is only used to set the (struct
tree).current field.

Delete both of those fields and the associated code which sets them.