Set struct pointers to NULL after free
None of these are strictly necessary -- that is, none of these pointers
are used after the free() statements.  However, it's good practice to
set them to NULL in case any new code is tempted to use (or free) them,
unless it's obvious that they're not going to be used again.

Check with:
    git grep -B1 -A9 "free([^)]*\."
    git grep -B1 -A9 "free([^)]*->"