pkg/debian: move into debian/compat-9/
In order to continue to build our Ubuntu 12.04 "precise" and Debian 8
"jessie" packages, we need to keep this deb packaging as-is.  However,
at some point we will need to modify the deb packaging for later
distros.  That will take place in a compat-10/ or compat-11/ directory.

This will complicate our package generation: we'll no longer have a
single tarsnap_X.Y.Z-r.debian.tar.gz; rather we'll have two or more
files with that identical name (as required by Debian) in different
subdirectories.  In other words, the website pkg/deb-src/ will be
rearranged similarly to pkg/deb/: there will be a ton of symlinks so
that each distro's codename points to the appropriate compat-X/
directory.