chunkify: Prohibit small-alphabet cycles
A chosen-plaintext attack can leak information about the chunking
parameters using "small-alphabet" inputs; if a file only contains
two distinct byte values then the only chunking parameters used to
chunk it are the prime p, multiplier alpha, and two coefficients
from cm[], which is a small enough space to be solvable.

This attack becomes infeasible when more coefficients in cm[] are
involved, so require a minimum of 8 distinct byte values (aka 8 or
more cm[] coefficients) in any cycle; if we find a cycle with fewer
distinct byte values, ignore it and continue the current chunk.

For more details on the attack, see
Alexeev, B., Percival, C., and Zhang, Y. X, Chunking Attacks on
File Backup Services using Content-Defined Chunking.  Unpublished
manuscript, March 2025.
which will be submitted shortly to the Cryptology ePrint Archive.